refactored enum

+++ PracticeGamestore/PracticeGamestore.API/Mappers/GameMappingExtensions.cs
@@ -30,7 +30,7 @@ public static class GameMappingExtensions
             Id = responseDto.Id,
             Name = responseDto.Name,
             Key = responseDto.Key,
-            AgeRating = responseDto.AgeRating,
+            AgeRating = (int)responseDto.AgeRating,
             Price = responseDto.Price,
             Description = responseDto.Description,
             Rating = responseDto.Rating,
